+++ app/server/controllers/ApplicationController.js
@@ -71,6 +71,8 @@ ApplicationController.prototype.getScopesAction = function(request, response) {
  * @param {String|Array} [request.query.exclude] The list of fields to exclude from returned applications. Ignored if
  * include is also specified.
  * @param {String} [request.query.query] Search query to search on application name
+ * @param {Number} [request.query.useSmartSearch=1] 1 to use a more advanced search mechanism, 0 to use a simple search
+ * based on a regular expression
  * @param {Number} [request.query.page=0] The expected page in pagination system
  * @param {Number} [request.query.limit=10] The maximum number of expected results
  * @param {String} [request.query.sortBy="name"] The application field To sort by (only "name" is available right now)
@@ -87,6 +89,7 @@ ApplicationController.prototype.getEntitiesAction = function(request, response, 
       include: {type: 'array<string>'},
       exclude: {type: 'array<string>'},
       query: {type: 'string'},
+      useSmartSearch: {type: 'number', in: [0, 1], default: 1},
       limit: {type: 'number', gt: 0},
       page: {type: 'number', gte: 0, default: 0},
       sortBy: {type: 'string', in: ['name'], default: 'name'},
@@ -104,7 +107,12 @@ ApplicationController.prototype.getEntitiesAction = function(request, response, 
   var filter = new ResourceFilter();
 
   // Add search query
-  if (params.query) filter.search('"' + params.query + '"');
+  if (params.query) {
+    if (params.useSmartSearch)
+      filter.search('"' + params.query + '"');
+    else
+      filter.regex('name', new RegExp(openVeoApi.util.escapeTextForRegExp(params.query), 'i'));
+  }
 
   provider.get(
     filter,