+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ## BREAKING CHANGES
 
 - The client tableContainer API has changed. "filterBy" property is now more consistent between the different types, "key" is the web service parameter and "getValue" property will give you the possibility to change the value before calling the web service. "param" and "filterWithChildren" properties don't exist anymore, you can achieve the same behaviour using the new "getValue" property. Note that filters of type "text" are no longer merged. Plugins will have to remove their "param" property, modify their "key" property and add a "getValue" property if they were using "param" property with values "date", "dateStart" or "startEnd", or if they were using the property "fillWithChildren". See documentation for more details
+- Plugins defining entities must now support the "useSmartSearch" in their getEntitiesAction method to pass web service end to end tests
 
 ## NEW FEATURES
 
 - The EntityService.getAllEntities function now accepts parameters
+- Web service endpoints GET /applications, GET /roles, GET /taxonomies, GET /groups and GET /users, now accept a "useSmartSearch" parameter which modifies the way search is made. If "useSmartSearch" is activated (this is the default), search will be made using the search mechanism of the storage. If "useSmartSearch" is deactivated it will search using a simple regular expression
 
 # 6.0.0 / 2018-10-17